code/kafka-ci360-connector/requirements.txt
configparser
confluent_kafka
requests
pyjwt
